requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.3
beautifulsoup4==4.11.2
bidict==0.23.1
blinker==1.8.0
Bootstrap-Flask==2.2.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
enum-compat==0.0.3
et-xmlfile==1.1.0
eventlet==0.36.1
Flask~=3.1.0
flask-babel==4.0.0
Flask-CKEditor==0.4.6
Flask-Dropzone==2.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
flask-mailman==1.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
instaloader==4.11
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pdfkit==1.0.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-slugify==8.0.4
python-socketio==5.11.2
pytz==2024.1
redis==5.0.4
requests==2.28.2
selenium==4.21.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.25
text-unidecode==1.3
trio==0.25.1
trio-websocket==0.11.1
twilio==9.1.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
visitor==0.1.3
webdriver-manager==4.0.1
Werkzeug~=3.1.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
